Rework DeckList component
