+++ frontend/components/navigation/Tabs/GroupsTab.tsx
@@ -1,5 +1,15 @@
+import { msg, t } from "@lingui/macro";
+import { useLingui } from "@lingui/react";
+import { useRouter } from "next/router";
+import { useRef } from "react";
+import { PlusSquare } from "react-feather";
+import { toast } from "react-toastify";
+import { mutate } from "swr";
+
 import { Group as GroupType } from "../../../types/Group";
-import DeckOverview from "../../deck/DeckList";
+import { authedFetch } from "../../../util/reauth";
+import DeckList from "../../deck/DeckList";
+import { InputField } from "../../form/InputField";
 
 interface GroupsTabProps {
 	/**
@@ -16,19 +26,74 @@ interface GroupsTabProps {
  * UI component for the GroupsTab
  */
 export const GroupsTab = ({ groups, className = "" }: GroupsTabProps) => {
+	const router = useRouter();
+	const { _ } = useLingui();
+
+	const groupNameInput = useRef<HTMLInputElement>(null);
+
 	return (
-		<div className={`${className}`}>
-			{groups
-				?.filter((group: GroupType) => !group.isDefault)
-				.map((group: GroupType) => {
-					return (
-						<DeckOverview
-							key={group.groupID}
-							group={group}
-						></DeckOverview>
-					);
-				})}
-			<DeckOverview />
+		<div className={`space-y-3 ${className}`}>
+			<div className="flex w-full items-center justify-between rounded-md bg-neutral-100 px-4 py-3">
+				<InputField
+					id={`groupNameInput`}
+					placeholder={_(msg`Create new Group`)}
+					inputFieldSize="xs"
+					className="w-full bg-transparent font-medium text-kiokuDarkBlue outline-none"
+					onKeyUp={(event) => {
+						if (event.key === "Enter") {
+							createGroup()
+								.then((result) => {})
+								.catch((error) => {});
+						}
+					}}
+					ref={groupNameInput}
+				/>
+				<PlusSquare
+					className="text-kiokuDarkBlue transition hover:scale-110 hover:cursor-pointer"
+					onClick={createGroup}
+				/>
+			</div>
+			<div className={`space-y-3 ${className}`}>
+				{groups
+					?.filter((group: GroupType) => !group.isDefault)
+					.map((group: GroupType) => {
+						return (
+							<div
+								className="hover:cursor-pointer"
+								onClick={() =>
+									router.push(`/group/${group.groupID}`)
+								}
+							>
+								<DeckList
+									header={group.groupName}
+									key={group.groupID}
+								/>
+							</div>
+						);
+					})}
+				<DeckList />
+			</div>
 		</div>
 	);
+
+	async function createGroup() {
+		if (!groupNameInput.current?.value) {
+			groupNameInput.current?.focus();
+			return;
+		}
+		const response = await authedFetch(`/api/groups`, {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+			},
+			body: JSON.stringify({ groupName: groupNameInput.current.value }),
+		});
+		if (response?.ok) {
+			groupNameInput.current.value = "";
+			toast.info(t`Group created!`, { toastId: "newGroupToast" });
+		} else {
+			toast.error("Error!", { toastId: "newGroupToast" });
+		}
+		mutate(`/api/groups`);
+	}
 };
